migrated to extension based registration of state handler implementations
