+++ bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm/META-INF/MANIFEST.MF
@@ -12,8 +12,8 @@ Require-Bundle: org.eclipse.core.resources,
  org.eclipse.emf.ecore,
  de.uka.ipd.sdq.identifier;bundle-version="5.0.0",
  org.palladiosimulator.pcm;bundle-version="5.0.0",
- org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel;bundle-version="0.1.0",
  org.palladiosimulator.dataflow.dictionary.characterized;bundle-version="5.1.0",
  org.palladiosimulator.dataflow.confidentiality.pcm.model.profile;bundle-version="5.1.0",
- org.palladiosimulator.dataflow.confidentiality.pcm.model;bundle-version="5.1.0"
+ org.palladiosimulator.dataflow.confidentiality.pcm.model;bundle-version="5.1.0",
+ org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel;bundle-version="0.1.0"
 Export-Package: org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm
