+++ bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm/plugin.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.statehandler_implementation">
+      <StateHandlerImplementation
+            class="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.AllocationStateHandler"
+            id="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.AllocationStateHandler"
+            name="AllocationStateHandler">
+      </StateHandlerImplementation>
+      <StateHandlerImplementation
+            class="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.AssemblyStateHandler"
+            id="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.AssemblyStateHandler"
+            name="AssemblyStateHandler">
+      </StateHandlerImplementation>
+      <StateHandlerImplementation
+            class="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.BranchStateHandler"
+            id="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.BranchStateHandler"
+            name="BranchStateHandler">
+      </StateHandlerImplementation>
+      <StateHandlerImplementation
+            class="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.CharacteristicsStateHandler"
+            id="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler.pcm.CharacteristicsStateHandler"
+            name="CharacteristicsStateHandler">
+      </StateHandlerImplementation>
+   </extension>
+
+</plugin>
