--- bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm/src/org/palladiosimulator/pcm/uncertainty/variation/UncertaintyVariationModel/gen/pcm/statespace/statehandler/StateHandlerFactory.java
@@ -1,7 +1,0 @@
-package org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm.statespace.statehandler;
-
-import UncertaintyVariationModel.VariationPoint;
-
-public interface StateHandlerFactory {
-    public StateHandler createFor(VariationPoint object);
-}
