+++ bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel.gen.pcm/src/org/palladiosimulator/pcm/uncertainty/variation/UncertaintyVariationModel/gen/pcm/statespace/statehandler/pcm/CharacteristicsStateHandler.java
@@ -13,6 +13,7 @@ import org.palladiosimulator.dataflow.dictionary.characterized.DataDictionaryCha
 import UncertaintyVariationModel.ValueCollection;
 import UncertaintyVariationModel.VariationDescription;
 import UncertaintyVariationModel.VariationPoint;
+import UncertaintyVariationModel.statehandler.GenericStateHandler;
 
 public class CharacteristicsStateHandler extends GenericStateHandler {
