+++ bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel/build.properties
@@ -4,7 +4,8 @@ bin.includes = .,\
                model/,\
                META-INF/,\
                plugin.xml,\
-               plugin.properties
+               plugin.properties,\
+               schema/statehandler_implementation.exsd
 jars.compile.order = .
 source.. = src/,\
            src-gen/
