+++ bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel/plugin.xml
@@ -5,6 +5,7 @@
 -->
 
 <plugin>
+   <extension-point id="statehandler_implementation" name="Statehandler Implementation" schema="schema/statehandler_implementation.exsd"/>
    <extension point="org.eclipse.emf.ecore.generated_package">
       <!-- @generated UncertaintyVariationModel -->
       <package
