+++ bundles/org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel/src/UncertaintyVariationModel/statehandler/StateHandlerFactory.java
@@ -1,0 +1,10 @@
+package UncertaintyVariationModel.statehandler;
+
+import UncertaintyVariationModel.VariationPoint;
+
+public interface StateHandlerFactory {
+    StateHandlerFactory INSTANCE = UncertaintyVariationModel.statehandler.impl.ConcreteStateHandlerFactory
+        .init();
+
+    public StateHandler createFor(VariationPoint object);
+}
