+++ features/org.palladiosimulator.pcm.uncertainty.variation.feature/feature.xml
@@ -28,10 +28,10 @@
       <import plugin="org.eclipse.ui.ide"/>
       <import plugin="de.uka.ipd.sdq.identifier" version="5.0.0" match="greaterOrEqual"/>
       <import plugin="org.palladiosimulator.pcm" version="5.0.0" match="greaterOrEqual"/>
-      <import plugin="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel" version="0.1.0" match="greaterOrEqual"/>
       <import plugin="org.palladiosimulator.dataflow.dictionary.characterized" version="5.1.0" match="greaterOrEqual"/>
       <import plugin="org.palladiosimulator.dataflow.confidentiality.pcm.model.profile" version="5.1.0" match="greaterOrEqual"/>
       <import plugin="org.palladiosimulator.dataflow.confidentiality.pcm.model" version="5.1.0" match="greaterOrEqual"/>
+      <import plugin="org.palladiosimulator.pcm.uncertainty.variation.UncertaintyVariationModel" version="0.1.0" match="greaterOrEqual"/>
       <import plugin="org.eclipse.osgi"/>
       <import plugin="org.eclipse.ui" version="3.118.100" match="greaterOrEqual"/>
       <import plugin="org.eclipse.e4.core.services" version="2.2.500" match="greaterOrEqual"/>
